Move tests for shiftable module into separate file
